+++ jobs/Frontend/src/types/APIResponse.ts
@@ -1,0 +1,8 @@
+import {Movie} from "@/types/Movie";
+
+export interface APIResponse {
+    page: number
+    results: Movie[]
+    total_pages: number
+    total_results: number
+}
